Create tensorize spectra file
