+++ tests/test_data_generators.py
@@ -5,9 +5,9 @@ import torch
 from matchms import Spectrum
 from ms2deepscore.train_new_model.data_generators import (DataGeneratorPytorch,
                                                           compute_validation_set,
-                                                          tensorize_spectra,
                                                           write_to_pickle,
                                                           load_generator_from_pickle)
+from ms2deepscore.tensorize_spectra import tensorize_spectra
 from ms2deepscore.SettingsMS2Deepscore import \
     GeneratorSettings, TensorizationSettings
 from ms2deepscore.train_new_model.spectrum_pair_selection import \
