+++ tests/test_siamese_spectra_model.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 from matchms import Spectrum
-from ms2deepscore.train_new_model.data_generators import (DataGeneratorPytorch,
-                                                          tensorize_spectra)
+from ms2deepscore.train_new_model.data_generators import (DataGeneratorPytorch)
+from ms2deepscore.tensorize_spectra import tensorize_spectra
 from ms2deepscore.models.SiameseSpectralModel import (SiameseSpectralModel,
                                                       train)
 from ms2deepscore.SettingsMS2Deepscore import GeneratorSettings, TensorizationSettings
